series

+++ app/Jobs/SuckSeriesJob.php
@@ -39,6 +39,10 @@ class SuckSeriesJob implements ShouldQueue
      */
     public function handle()
     {
+        $is_recent = Serie::where('id', $this->id)
+        ->where('updated_at', '>', Carbon::now()->subHours(5)->toDateTimeString())
+        ->first();
+        if($is_recent) return;
         if($this->isWithRecommendation){
             $series = json_decode(file_get_contents('https://api.themoviedb.org/3/tv/'.$this->id.'?api_key='.config('constants.api_key').'&language=en&append_to_response=recommendations,reviews'), true);
             Series_recommendation::where(['series_id' => $this->id])->delete();
@@ -127,11 +131,6 @@ class SuckSeriesJob implements ShouldQueue
                 );
             }
         }else{
-            $is_recent = Serie::where('id', $this->id)
-            ->where('updated_at', '>', Carbon::now()->subHours(5)->toDateTimeString())
-            ->first();
-            if($is_recent) return;
-            
             $series = json_decode(file_get_contents('https://api.themoviedb.org/3/tv/'.$this->id.'?api_key='.config('constants.api_key').'&language=en&append_to_response=reviews'), true);
             $series_tr = json_decode(file_get_contents('https://api.themoviedb.org/3/tv/'.$this->id.'?api_key='.config('constants.api_key').'&language=tr&append_to_response=reviews'), true);
             $series_hu = json_decode(file_get_contents('https://api.themoviedb.org/3/tv/'.$this->id.'?api_key='.config('constants.api_key').'&language=hu&append_to_response=reviews'), true);
